+++ src/application/Position.java
@@ -1,31 +1,31 @@
 package application;
 
 public class Position implements Comparable<Position>{
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Position(int x, int y) {
+	public Position(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getx() {
+	public double getx() {
 		return x;
 	}
 	
-	public int gety() {
+	public double gety() {
 		return y;
 	}
 	
-	public void setPos(int x, int y) {
+	public void setPos(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
 	@Override
 	public int compareTo(Position o) {
-		int pos_o = o.gety();
-		return this.y-pos_o;
+		double pos_o = o.gety();
+		return (int)(this.y-pos_o);
 	}
